Split field package into value package
